Changed Id type to Long Extended EditExpenseState dataclass for future changes

+++ app/src/main/java/com/example/expensetracker/data/model/Expense.kt
@@ -7,5 +7,5 @@ import androidx.room.PrimaryKey
 data class Expense(
     val name: String,
     val amount: Long,
-    @PrimaryKey(autoGenerate = true) val id: Int = 0
+    @PrimaryKey(autoGenerate = true) val id: Long = 0
 )
